% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{alignment}
\alias{alignment}
\title{Alignment of signals of the dataset through the CluPA algorithm provided by the speaq package.}
\usage{
alignment(dataset, buck_step)
}
\arguments{
\item{dataset}{The 1D NMR dataset where to align signals}

\item{buck_step}{The bucketing of the dataset (e.g. 0.001)}
}
\value{
aligneddataset The dataset with the signals aligned
}
\description{
Alignment of signals of the dataset through the CluPA algorithm provided by the speaq package.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
aligned_data=alignment(imported_data$dataset,imported_data$buck_step)
}
