% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autorun.R
\name{automatic_profiling}
\alias{automatic_profiling}
\title{Automatic quantification of signals for all experiments using the information located in the ROI patterns file.}
\usage{
automatic_profiling(imported_data, ROI_data, optimization = TRUE,
  spectra_to_profile = NA)
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{ROI_data}{ROIs data.}

\item{optimization}{By default TRUE. If TRUE, profiling quality is maximized through the analysis og signals parameters, with the tradeoff of additional computing time.}

\item{spectra_to_profile}{By default NA, and all spectra are profiled. If a vector of spectra is provided, profiling is limited to this vector.}
}
\value{
List with final_output (with metabolite signal relative concentrations and quality indicators) and reproducibility_data (with the necessary data to reproduce the profiling performed).
}
\description{
Automatic quantification of signals for all experiments using the information located in the ROI patterns file.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
# Not run:
# profiling_data=automatic_profiling(imported_data,imported_data$ROI_data)
}
