% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_profiling_improv.R
\name{automatic_profiling_improv}
\alias{automatic_profiling_improv}
\title{Improvement of automatic profiling using the inforamtion collected on a previous implementation.}
\usage{
automatic_profiling_improv(imported_data, final_output,
  reproducibility_data, ROI_data,
  improvement_option = "reimplementation", level = "outliers")
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{reproducibility_data}{List with necessary information to load quantifications on the Shiny GUI.}

\item{ROI_data}{ROIs data.}

\item{improvement_option}{If "correction", quantifications are updated taking into account the predicted signal parameters. If "reimplemetation", profiling is repeated using the prediction information.}

\item{level}{How extensive should be the improvement? If "all", all quantifications are changed. If "outliers", quantifications whoss signal parameters behave as outliers are changed (please take into account that only the quantifications will be updated). If a number is introduced, the quantifications with a higher fitting error than the number specified are repeated.}
}
\value{
List with final_output (with metabolite signal relative concentrations and quality indicators) and reproducibility_data (with the necessary data to reproduce the profiling performed).
}
\description{
Improvement of automatic profiling using the inforamtion collected on a previous implementation.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
# Not run:
# load(file.path(system.file(package = "rDolphin"),"extdata","MTBLS242_subset_profiling_data.RData"))
# profiling_data_2=automatic_profiling_improv(imported_data,profiling_data$final_output,profiling_data$reproducibility_data,imported_data$ROI_data)
}
