% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autorun.R
\name{autorun}
\alias{autorun}
\title{Automatic quantification of signals for all experiments using the information located in the ROI patterns file.}
\usage{
autorun(imported_data, finaloutput, useful_data, ROI_data, ROI_separator)
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{finaloutput}{List with quantifications and indicators of quality of quantification.}

\item{useful_data}{List with necessary information to load quantifications on the Shiny GUI.}

\item{ROI_data}{ROIs data.}

\item{ROI_separator}{ROI separator.}
}
\value{
List with updated finaloutput and useful_data variables.
}
\description{
Automatic quantification of signals for all experiments using the information located in the ROI patterns file.
}
\examples{
setwd(paste(system.file(package = "Dolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
quantification_variables=autorun(imported_data,imported_data$finaloutput,imported_data$useful_data,imported_data$ROI_data,imported_data$ROI_separator)
}
