% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autorun_model_spectrum.R
\name{autorun_model_spectrum}
\alias{autorun_model_spectrum}
\title{Automatic quantification in the model spectrum of the dataset using the information located in the ROI patterns file, with associated p values for every bin. The reulting information can be useful to analyze possible new ROIs to profile or modification of existing ones.}
\usage{
autorun_model_spectrum(imported_data)
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}
}
\value{
plotly figure with performed fitting for every ROI and p value for every bin
}
\description{
Automatic quantification in the model spectrum of the dataset using the information located in the ROI patterns file, with associated p values for every bin. The reulting information can be useful to analyze possible new ROIs to profile or modification of existing ones.
}
\examples{
setwd(paste(system.file(package = "Dolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
model_spectrum_plot=autorun_model_spectrum(imported_data)
}
