% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_variables.R
\name{fitting_variables}
\alias{fitting_variables}
\title{Preparation of parameters involved with the program. A list of manual changes to parameters can be provided.}
\usage{
fitting_variables(parameter_change = list())
}
\arguments{
\item{parameter_change}{List of parameters ot change, if desired.}
}
\value{
Preparation of parameters involved with the program
}
\description{
Preparation of parameters involved with the program. A list of manual changes to parameters can be provided.
}
\examples{
program_parameters=fitting_variables()
program_parameters=fitting_variables(list(fitting_maxiterrep=1,fitting_error_analysis_limit=0.05))
}
