% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification_tool.R
\name{identification_tool}
\alias{identification_tool}
\title{STOCSY or RANSY identification tool.}
\usage{
identification_tool(dataset, ppm, driver_peak_edges, method)
}
\arguments{
\item{dataset}{Dataset}

\item{ppm}{ppm axis}

\item{driver_peak_edges}{Left and right edges of the driver peak to correlate with the rest of spectrum.}

\item{method}{Method ('pearson','spearman','ransy').}
}
\value{
interactive plot
}
\description{
STOCSY or RANSY identification tool.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
identification_tool(imported_data$dataset,imported_data$ppm,c(1,0.995),'spearman')
}
