% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import of variables stored in the parameters file and of the dataset to quantify}
\usage{
import_data(parameters_path)
}
\arguments{
\item{parameters_path}{CSV file woth input paths and desired pre-processing steps}
}
\value{
Imported data of experiment
}
\description{
Import of variables stored in the parameters file and of the dataset to quantify
}
\examples{
setwd(paste(system.file(package = "Dolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
}
