% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_error.R
\name{integration_error}
\alias{integration_error}
\title{Helper function to generate ROI data. Not to be used in console.}
\usage{
integration_error(ROI_data, useful_data, final_output, ind)
}
\arguments{
\item{ROI_data}{ROI_data}

\item{useful_data}{useful_data}

\item{final_output}{final_output}

\item{ind}{ind}
}
\value{
dummy
}
\description{
Helper function to generate ROI data. Not to be used in console.
}
