% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_quantification.R
\name{load_quantification}
\alias{load_quantification}
\title{Load quantification}
\usage{
load_quantification(reproducibility_data, imported_data, final_output, info,
  ROI_data)
}
\arguments{
\item{reproducibility_data}{List with necessary information to load quantifications on the Shiny GUI.}

\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{info}{List with 'row' and 'column' indicating spectrum and signal to load.}

\item{ROI_data}{ROIs data}
}
\value{
Loaded plot, signals parameters and quality of fitting parameters of quantifications
}
\description{
Load quantification
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
resulting_data=individual_profiling(imported_data,imported_data$final_output,c(1,4),imported_data$ROI_data[3:4,],imported_data$reproducibility_data)
loaded_quantification=load_quantification(resulting_data$reproducibility_data,imported_data,resulting_data$final_output,list(row=1,col=4),imported_data$ROI_data)
}
