% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianplot.R
\name{medianplot}
\alias{medianplot}
\title{Creation of plotly figure with the median spectrum for each group of spectra}
\usage{
medianplot(imported_data)
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}
}
\value{
Plotly figure with the median spectrum for each group of spectra
}
\description{
Creation of plotly figure with the median spectrum for each group of spectra
}
\examples{
setwd(paste(system.file(package = "Dolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
median_plot=medianplot(imported_data)
}
