% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_values}
\alias{p_values}
\title{Univariate analysis}
\usage{
p_values(dataset, metadata)
}
\arguments{
\item{dataset}{dataset of bins or quantifications.}

\item{metadata}{Metadata to be used during the univariate analysis.}
}
\value{
p values from dataset of bins or quantificaitons
}
\description{
Univariate analysis
}
\examples{
setwd(paste(system.file(package = "Dolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
pval=p_values(imported_data$dataset,imported_data$Metadata)
}
