% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_quant.R
\name{remove_quant}
\alias{remove_quant}
\title{Import of variables stored in the parameters file and of the dataset to quantify}
\usage{
remove_quant(info, imported_data, final_output)
}
\arguments{
\item{info}{Experiment and signal where to remove information}

\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{final_output}{List with quantifications and indicators of quality of quantification.}
}
\value{
Imported data of experiment
}
\description{
Import of variables stored in the parameters file and of the dataset to quantify
}
