% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roifunc.r
\name{roifunc}
\alias{roifunc}
\title{Helper function to generate ROI data}
\usage{
roifunc(ROI_data, ROI_separator, Metadata, Experiments)
}
\arguments{
\item{ROI_data}{ROIs data}

\item{Metadata}{Metadata}

\item{Experiments}{Experiments}

\item{datapath}{Type of valdiation to perform (1: fitting error, 2: signal area ratio, 3: chemical shift, 4: half bandwidth, 5: outliers, 6: relative intensity of signals of same metabolite)}

\item{ROI_data}{ROI separator}
}
\value{
dummy
}
\description{
Helper function to generate ROI data
}
