% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_roi_testing.R
\name{save_roi_testing}
\alias{save_roi_testing}
\title{Saving of results in interface in individual quantification}
\usage{
save_roi_testing(provisional_data, imported_data, finaloutput, useful_data)
}
\arguments{
\item{provisional_data}{List with elements to save.}

\item{imported_data}{Imported_data.}

\item{finaloutput}{List with quantifications and indicators of quality of quantification.}

\item{useful_data}{List with necessary information to load quantifications on the Shiny GUI.}
}
\value{
List with updated finaloutput and useful_data variables.
}
\description{
Saving of results in interface in individual quantification
}
