% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signals_int.R
\name{signals_int}
\alias{signals_int}
\title{Helper function to edit line shape fitting}
\usage{
signals_int(imported_data, final_output, spectrum_index, signals_introduce,
  ROI_profile)
}
\arguments{
\item{imported_data}{imported data}

\item{final_output}{final_output}

\item{spectrum_index}{spectrum_index}

\item{signals_introduce}{signals_introduce}

\item{ROI_profile}{ROI_profile}
}
\value{
provisional_data
}
\description{
Helper function to edit line shape fitting
}
