% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOCSY.R
\name{STOCSY}
\alias{STOCSY}
\title{Automatic quantification of signals for all experiments using the information located in the ROI patterns file.}
\usage{
STOCSY(dataset, ppm, limits, method)
}
\arguments{
\item{dataset}{Dataset}

\item{ppm}{ppm}

\item{limits}{Limits}

\item{method}{Correlation method.}

\item{ROI_separator}{ROI separator.}
}
\value{
STOCSY plot
}
\description{
Automatic quantification of signals for all experiments using the information located in the ROI patterns file.
}
