% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twod_data.R
\name{twod_data}
\alias{twod_data}
\title{Loading of 2D NMR spectrum, based on a function in rNMR.}
\usage{
twod_data(path)
}
\arguments{
\item{path}{2D path}
}
\value{
2D data
}
\description{
Loading of 2D NMR spectrum, based on a function in rNMR.
}
