% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_plot.R
\name{type_plot}
\alias{type_plot}
\title{Type of analysis plot}
\usage{
type_plot(imported_data, ROI_limits, rows_selected, medianplot, clusterplot)
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{ROI_limits}{Range of plot}

\item{rows_selected}{Rows selected in GUi table.}

\item{medianplot}{Meidan plot}

\item{clusterplot}{Cluster plot}
}
\value{
Plot
}
\description{
Type of analysis plot
}
