% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autorun.R
\name{autorun}
\alias{autorun}
\title{Automatic quantification of signals for all experiments using the information located in the ROI patterns file.}
\usage{
autorun(imported_data, final_output, useful_data, ROI_data)
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{useful_data}{List with necessary information to load quantifications on the Shiny GUI.}

\item{ROI_data}{ROIs data.}
}
\value{
List with updated final_output and useful_data variables.
}
\description{
Automatic quantification of signals for all experiments using the information located in the ROI patterns file.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
# Not run:
# profiling_data=autorun(imported_data,imported_data$final_output,imported_data$useful_data,imported_data$ROI_data)
}
