% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperimport.R
\name{helperimport}
\alias{helperimport}
\title{Helper function to import data in GUI. Not to be used in console.}
\usage{
helperimport(datapath, dummy)
}
\arguments{
\item{dummy}{}
}
\value{
dummy
}
\description{
Helper function to import data in GUI. Not to be used in console.
}
