% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renew_imported_data.R
\name{renew_imported_data}
\alias{renew_imported_data}
\title{Adaptation of variables of imported data to changes applied to the ROI profiles}
\usage{
renew_imported_data(imported_data)
}
\arguments{
\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}
}
\value{
Changed imported data
}
\description{
Adaptation of variables of imported data to changes applied to the ROI profiles
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
imported_data$ROI_data=imported_data$ROI_data[-1,]
imported_data=renew_imported_data(imported_data)
}
