% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roifunc.r
\name{roifunc}
\alias{roifunc}
\title{Helper function to generate ROI data. Not to be used in console.}
\usage{
roifunc(ROI_data, Metadata, Experiments)
}
\arguments{
\item{ROI_data}{ROIs data}

\item{Metadata}{Metadata}

\item{Experiments}{Experiments}
}
\value{
dummy
}
\description{
Helper function to generate ROI data. Not to be used in console.
}
