% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification_tool.R
\name{identification_tool}
\alias{identification_tool}
\title{STOCSY or RANSY interactive plot.}
\usage{
identification_tool(dataset, ppm, limits, method)
}
\arguments{
\item{dataset}{Dataset}

\item{ppm}{ppm}

\item{limits}{Left and right edges of the region to correalte with the rest of spectrum.}

\item{method}{Correlation method ('pearson','spearman').}
}
\value{
STOCSY or RANSY interactive plot
}
\description{
STOCSY or RANSY interactive plot.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
identification_tool(imported_data$dataset,imported_data$ppm,c(2.14,2.12),method='spearman')
}
