% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOCSY.R
\name{STOCSY}
\alias{STOCSY}
\title{Automatic quantification of signals for all experiments using the information located in the ROI patterns file.}
\usage{
STOCSY(dataset, ppm, limits, method)
}
\arguments{
\item{dataset}{Dataset}

\item{ppm}{ppm}

\item{limits}{Left and right edges of the region to correalte with the rest of spectrum.}

\item{method}{Correlation method ('pearson','spearman').}
}
\value{
STOCSY interactive plot
}
\description{
Automatic quantification of signals for all experiments using the information located in the ROI patterns file.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
STOCSY(imported_data$dataset,imported_data$ppm,c(1,0.995),'spearman')
}
