% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_analysis_plot.R
\name{type_analysis_plot}
\alias{type_analysis_plot}
\title{Type of analysis plot}
\usage{
type_analysis_plot(data, final_output, imported_data, type = c("boxplot",
  "pca", "dendrogram_heatmap"))
}
\arguments{
\item{data}{Data to be used to get the analysis plot}

\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{type}{Kind of plot wanted ('boxplot','pca','dendrogram_heatmap')}
}
\value{
Analysis plot
}
\description{
Type of analysis plot
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
load("MTBLS242_subset_profiling_data.RData")
type_analysis_plot(profiling_data$final_output$quantification,profiling_data$final_output,imported_data,'boxplot')
}
