% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Estimation of quality indicators, with signal to total area ratio, fitting error, and difference between expected and found signal parameter values.}
\usage{
validation(final_output, validation_type, alarm_matrix = NA)
}
\arguments{
\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{validation_type}{Type of validation to perform (1: fitting error, 2: signal total area ratio, 3: difference expected-obtained chemical shift, 4: difference expected-obtained half bandwidth, 5: difference expected-obtained intensity)}

\item{alarm_matrix}{Optional list of previous indicators to update. By default NA.}
}
\value{
List with estimated indicators, matrix to show in GUI, and breaks and colors to prepare a different color for each cell of the matrix.
}
\description{
Estimation of quality indicators, with signal to total area ratio, fitting error, and difference between expected and found signal parameter values.
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
load("MTBLS242_subset_profiling_data.RData")
validation_data=validation(profiling_data$final_output,5)
}
