% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Creation of matrix to validate quantifications, with information of difference with predicted shift, signal to total area ratio, fitting error, and difference with expected relative intensity}
\usage{
validation(final_output, alarmmatrix, validation_type)
}
\arguments{
\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{alarmmatrix}{List with quantifications and indicators of quality of quantification.}

\item{validation_type}{Type of valdiation to perform (1: fitting error, 2: signal area ratio, 3: chemical shift, 4: half bandwidth, 5: outliers, 6: relative intensity of signals of same metabolite)}
}
\value{
Matrix with data required
}
\description{
Creation of matrix to validate quantifications, with information of difference with predicted shift, signal to total area ratio, fitting error, and difference with expected relative intensity
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
load("MTBLS242_subset_profiling_data.RData")
validation_data=validation(profiling_data$final_output,profiling_data$alarmmatrix,5)
}
