% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation}
\alias{validation}
\title{Creation of matrix to validate quantifications, with information of difference with predicted shift, signal to total area ratio, fitting error, and difference with expected relative intensity}
\usage{
validation(final_output, validation_type, ROI_data, metadata)
}
\arguments{
\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{validation_type}{Type of valdiation to perform (1: fitting error, 2: signal area ratio, 3: chemical shift, 4: half bandwidth, 5: outliers, 6: relative intensity of signals of same metabolite)}

\item{ROI_data}{ROIs data}

\item{metadata}{Dataset metadata}
}
\value{
Matrix with data required
}
\description{
Creation of matrix to validate quantifications, with information of difference with predicted shift, signal to total area ratio, fitting error, and difference with expected relative intensity
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
imported_data=import_data("Parameters_MTBLS242_15spectra_5groups.csv")
quantification_variables=autorun(imported_data,imported_data$final_output,imported_data$useful_data)
validation_data=validation(quantification_variables$final_output,5,imported_data$ROI_data,imported_data$Metadata)
DT::datatable(round(validation_data$alarmmatrix,4),selection = list(mode = 'single', target = 'cell')) \%>\% formatStyle(colnames(validation_data$alarmmatrix), backgroundColor = styleInterval(validation_data$brks, validation_data$clrs))
}
