% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plots.R
\name{write_plots}
\alias{write_plots}
\title{Import of variables stored in the parameters file and of the dataset to quantify}
\usage{
write_plots(path, final_output, imported_data, useful_data)
}
\arguments{
\item{path}{Path where plots are stored inside a created 'plots' folder}

\item{final_output}{List with quantifications and indicators of quality of quantification.}

\item{imported_data}{List with typical elements necessary to perform quantification of ROIs.}

\item{useful_data}{List with necessary information to load quantifications on the Shiny GUI.}
}
\value{
Plots in pdf files
}
\description{
Import of variables stored in the parameters file and of the dataset to quantify
}
\examples{
setwd(paste(system.file(package = "rDolphin"),"extdata",sep='/'))
load("MTBLS242_subset_example.RData")
#Not run:
#write_plots('quantification_plots',quantification_variables$final_output,imported_data,quantification_variables$useful_data)
}
